+++ core-mpi.py
@@ -5,31 +5,34 @@
 # parallel code skelton: work in progress
 #
 
-import numpy as np
-import pandas as pd
 import sys
 import time
+
 import h5py  # to read hdf5 files
+import numpy as np
+import pandas as pd
 import yaml  # to parse yaml config files
 from astropy.cosmology import FlatLambdaCDM  # cosmology model
-from scipy.interpolate import CubicSpline   # for interpolations
-# from sklearn.neighbors import BallTree      # for nearest neighbours
-from kdtreecode import BallTree             # for nearest neighbours
-# for loading the catalogs
-from reading_data_shape_redshift_catalog import (
-    reading_lens_params, reading_data_sources
-)
-# for calculating delta-sigma
-from calc_tngt_shear import get_lens_constants, calculate_dsigma_increments
-
 from mpi4py import MPI
+from scipy.interpolate import CubicSpline  # for interpolations
+
+# for calculating delta-sigma
+from src.calc_tngt_shear import calculate_dsigma_increments, get_lens_constants
+
+# from sklearn.neighbors import BallTree      # for nearest neighbours
+from src.kdtreecode import BallTree  # for nearest neighbours
+
+# for loading the catalogs
+from src.reading_data_shape_redshift_catalog import (
+    reading_data_sources,
+    reading_lens_params,
+)
 
 # define the function to run pipline.
 # input: config filename
 
 
 def run_pipeline(config_fname):
-
     comm = MPI.COMM_WORLD
     size = comm.Get_size()
     rank = comm.Get_rank()  # rank of the process
@@ -38,72 +41,70 @@ def run_pipeline(config_fname):
     # reading the config file
     #
     sys.stderr.write("Reading the config file...\n")
-    with open(config_fname, 'r') as file:
+    with open(config_fname, "r") as file:
         inputs = yaml.safe_load(file)
 
     #
     # creating a flat-lcdm model using the given cosmology parameters.
     #
     cm = FlatLambdaCDM(
-        H0=inputs['cosmology']['H0'],
-        Om0=inputs['cosmology']['OmegaMatter']
+        H0=inputs["cosmology"]["H0"], Om0=inputs["cosmology"]["OmegaMatter"]
     )  # defaults: Tcmb = 0K, Ob0 = None
 
     # redshift range
-    z_min = inputs['z_min']
-    z_max = inputs['z_max']
+    z_min = inputs["z_min"]
+    z_max = inputs["z_max"]
 
     #
     # calculate and interpolate the comoving distances for future use
     #
-    z = np.linspace(
-        max(z_min - 0.1, 0.),
-        min(z_max + 0.1, 10.),
-        inputs['z_bins']
-    )
+    z = np.linspace(max(z_min - 0.1, 0.0),
+                    min(z_max + 0.1, 10.0), inputs["z_bins"])
     xz = cm.comoving_distance(z)
 
     # spline object for comoving distance calculation
     comoving_distance = CubicSpline(z, xz)  # can be accesses like a function
 
     # parameters for binning neighbour distance
-    r_min = inputs['r_min']  # lower distance bound
-    r_max = inputs['r_max']  # upper distance bound
-    r_bins = inputs['r_bins']  # number of bins
+    r_min = inputs["r_min"]  # lower distance bound
+    r_max = inputs["r_max"]  # upper distance bound
+    r_bins = inputs["r_bins"]  # number of bins
 
     #
     # find the angular seperation for nn search
     #
     # comoving distance corresponding to minimum redshift
     cmdist = comoving_distance(z_min)
-    theta_max = r_max / cmdist          # maximum angular seperation
+    theta_max = r_max / cmdist  # maximum angular seperation
 
     #
     # read files: lenses catalog -> ra, dec, redshift
     #
     sys.stderr.write("Reading the lens catalog...\n")
-    lens_fname = inputs['files']['lens_file']  # lenses filename
+    lens_fname = inputs["files"]["lens_file"]  # lenses filename
 
     # read the lens data into a pandas.DataFrame object, having features including
     # coadd_object_id, ra, dec, zredmagic and lum_z
     # NOTE: ra and dec must be in radians
-    lenses = pd.DataFrame(reading_lens_params(
-        lens_fname, z_min, z_max, inputs['frac_bright']))
+    lenses = pd.DataFrame(
+        reading_lens_params(lens_fname, z_min, z_max, inputs["frac_bright"])
+    )
     lenses = lenses.dropna()  # dropping the nan
 
     # precalculate the lens constants and comoving dist
     lconst, lcmdist = get_lens_constants(lenses, comoving_distance)
-    lenses['const'] = lconst  # lens constants
-    lenses['cdist'] = lcmdist  # comoving distances to the lenses
+    lenses["const"] = lconst  # lens constants
+    lenses["cdist"] = lcmdist  # comoving distances to the lenses
 
     #
     # create a ball-tree using the lens positions for efficent neighbor search
     #
     sys.stderr.write("Creating the lens tree...\n")
-    lens_bt = BallTree(data=lenses[['dec', 'ra']].to_numpy(),
-                       # leaf_size = 20,
-                       # metric    = 'haversine' # metric on a spherical surface
-                       )
+    lens_bt = BallTree(
+        data=lenses[["dec", "ra"]].to_numpy(),
+        # leaf_size = 20,
+        # metric    = 'haversine' # metric on a spherical surface
+    )
     # lens_bt = BallTree( lenses[['dec', 'ra']].to_numpy(),
     #                )
 
@@ -112,13 +113,13 @@ def run_pipeline(config_fname):
     #
     sys.stderr.write("Creating source file objects...\n")
     # source shape data
-    srcs_file = h5py.File(inputs['files']['src_shape_file'], 'r')
+    srcs_file = h5py.File(inputs["files"]["src_shape_file"], "r")
     srcz_file = h5py.File(
-        inputs['files']['src_redshift_file'], 'r')  # source redshifts
+        inputs["files"]["src_redshift_file"], "r")  # source redshifts
 
     # size of the sources
     src_size = srcs_file["catalog"]["unsheared"]["e_1"].shape[0]
-    chunk_size = inputs['chunk_size']   # size of each sub catalog
+    chunk_size = inputs["chunk_size"]  # size of each sub catalog
 
     #
     # dividing the jonb among the processes
@@ -141,14 +142,14 @@ def run_pipeline(config_fname):
     dsigma_num_cross = np.zeros(r_bins - 1)
 
     # calculate the bin edges TODO
-    r_edges = np.logspace(np.log10(r_min), np.log10(
-        r_max), r_bins)  # log space bin edges
+    r_edges = np.logspace(
+        np.log10(r_min), np.log10(r_max), r_bins
+    )  # log space bin edges
 
     # nnDB   = [] # a database for the holding the source chunks and the neighbour data
     # dsigma = [] # to store the delta-sigma values (TODO: check this)
     sys.stderr.write("Starting mainloop...\n")
     for i in range(chunk_start, chunk_stop):
-
         # load a subset of sources
         start = i * chunk_size
         stop = start + chunk_size
@@ -156,10 +157,11 @@ def run_pipeline(config_fname):
         src_i = pd.DataFrame(reading_data_sources(
             srcs_file, srcz_file, start, stop))
         src_i = src_i.dropna()  # dropping the nan
-        src_i['cdist_mean'] = comoving_distance(
-            src_i['zmean_sof'])  # using mean redshift
-        src_i['cdist_mc'] = comoving_distance(
-            src_i['zmc_sof'])   # using mc redshift
+        src_i["cdist_mean"] = comoving_distance(
+            src_i["zmean_sof"]
+        )  # using mean redshift
+        src_i["cdist_mc"] = comoving_distance(
+            src_i["zmc_sof"])  # using mc redshift
 
         # find the nearest neighbours using the maximum radius
         sys.stderr.write("Searching for neighbours...\n")
@@ -169,8 +171,7 @@ def run_pipeline(config_fname):
         #                                   return_distance = True
         #                                )
 
-        nnid = lens_bt.query_radius(src_i[['dec', 'ra']].to_numpy(),
-                                    theta_max)
+        nnid = lens_bt.query_radius(src_i[["dec", "ra"]].to_numpy(), theta_max)
         sys.stderr.write(f"Completed in {time.time() - __t0:,} sec\n")
 
         # NOTE 1: `nnid` and `dist` are arrays of arrays so that, each sub-array
@@ -195,7 +196,8 @@ def run_pipeline(config_fname):
         # src_i, lenses, nnid, dist, r_edges
         # )
         delta_num, delta_num_cross, delta_den = calculate_dsigma_increments(
-            src_i, lenses, nnid, r_edges)
+            src_i, lenses, nnid, r_edges
+        )
         sys.stderr.write(f"Completed in {time.time() - __t0:,} sec\n")
 
         dsigma_num = dsigma_num + delta_num
@@ -212,9 +214,9 @@ def run_pipeline(config_fname):
     if rank > 0:
         # send data from every other process, except 0
         # FIXME is the `tag` can be any id
-        comm.Send(dsigma_num,       dest=0, tag=14)
+        comm.Send(dsigma_num, dest=0, tag=14)
         comm.Send(dsigma_num_cross, dest=0, tag=15)
-        comm.Send(denom,            dest=0, tag=16)
+        comm.Send(denom, dest=0, tag=16)
     else:
         # reciev from all others at 0
         for __proc in range(1, size):
@@ -242,16 +244,18 @@ def run_pipeline(config_fname):
         pd.DataFrame(
             {
                 # bin centers (linear)
-                'r_center': 0.5*(r_edges[1:] + r_edges[:-1]),
-                'dsigma': dsigma,
-                'dsigma_cross': dsigma_cross,
-            }).to_csv(
-            inputs['files']['output'],  # output filename
+                "r_center": 0.5 * (r_edges[1:] + r_edges[:-1]),
+                "dsigma": dsigma,
+                "dsigma_cross": dsigma_cross,
+            }
+        ).to_csv(
+            inputs["files"]["output"],  # output filename
             index=False,  # do not write the indices to the file
         )
 
         sys.stderr.write("The end...\n")
     return
+
 
 #
 # main function: parse the arguments and run the pipline
@@ -259,18 +263,18 @@ def run_pipeline(config_fname):
 
 
 def mainloop():
-
     import argparse
 
     # creating the argument parser
     parser = argparse.ArgumentParser(
-        description="Density profile calculations using weak lensing")
+        description="Density profile calculations using weak lensing"
+    )
     parser.add_argument(
-        'config',
-        metavar='file',
+        "config",
+        metavar="file",
         type=str,
-        nargs='?',
-        help='path to the configuration (yaml) file'
+        nargs="?",
+        help="path to the configuration (yaml) file",
     )
 
     # parsing the arguments. if a correct path to a config file given, run the pipeline
@@ -281,5 +285,5 @@ def mainloop():
     return
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     mainloop()
